requirements.txt
colorama==0.4.6
pillow==11.2.1
qrcode==8.2
